pyproject.toml
[tool.poetry]
name = "model-deployments"
version = "0.1.0"
description = ""
authors = ["Bobby Donchev <contact@donchev.is>", "Elena Cramer <elena.cramer@neurocode.io>"]

[tool.poetry.dependencies]
python = "^3.8"
azure-functions = "^1.4.0"
onnxruntime = "^1.5.2"
transformers = "^3.4.0"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.2"
torch = "^1.7.0"
onnx = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py
)
'''